pyproject.toml
[project]
name = "claude-latex-zotero-course-template"
version = "0.1.0"
description = "Course template with Beamer lectures, Jupyter notebooks, and Zotero integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymupdf",
    "beautifulsoup4",
    "markdownify",
    "pyyaml",
    "requests>=2.32.4",
    "docling>=2.42.0",
    "chromadb>=1.0.15",
    "sentence-transformers>=5.0.0",
    "nbformat>=5.10.4",
    "nbconvert>=7.17.0",
    "ipykernel>=7.2.0",
    "pytest>=8.0",
    "pypdfium2>=4.30.0",
    "playwright>=1.58.0",
    "transformer_lens>=2.0.0",
    "einops>=0.7.0",
    "circuitsvis>=1.0.0",
    "plotly>=5.0.0",
    "scikit-learn>=1.0.0",
    "jaxtyping>=0.2.0",
    "typeguard>=4.0.0",
    "huggingface_hub>=0.20.0",
    "matplotlib>=3.0.0",
]
